organise all constants use for tests between helpers and constants
