+++ tests/Helpers/ERC725Account.spec.ts
@@ -4,7 +4,7 @@ import { ethers } from "hardhat";
 
 import { ERC725Account, ERC725Account__factory } from "../../types";
 import { getRandomAddresses, generateKeysAndValues } from "../utils/helpers";
-import { BasicUPSetup_Schema } from "../utils/lsp2schema";
+import { BasicUPSetup_Schema } from "../utils/constants";
 
 describe("ERC725 Account", () => {
   let accounts: SignerWithAddress[];
